+++ src/main/java/controllers/tipoatracciones/TipoAtraccionesDeleteServlet.java
@@ -6,6 +6,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import model.PerfilUsuario;
 import services.TipoAtraccionService;
 import services.UsuarioService;
 
@@ -24,8 +25,13 @@ public class TipoAtraccionesDeleteServlet extends HttpServlet implements Servlet
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        Integer id = Integer.parseInt(req.getParameter("id"));
-        tipoAtraccionService.delete(id);
-        resp.sendRedirect("/tierramedia/tipos/lista.adm");
+        PerfilUsuario logedUser = (PerfilUsuario) req.getSession().getAttribute("user");
+        if (logedUser != null && logedUser.isAdmin()) {
+            Integer id = Integer.parseInt(req.getParameter("id"));
+            tipoAtraccionService.delete(id);
+            resp.sendRedirect("/tierramedia/tipos/lista.adm");
+        } else {
+            resp.sendRedirect("/tierramedia/welcome");
+        }
     }
 }
