+++ src/main/java/controllers/user/DeleteUsuarioServlet.java
@@ -8,6 +8,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import model.PerfilUsuario;
 import services.UsuarioService;
 
 @WebServlet("/usuario/delete.adm")
@@ -23,8 +24,13 @@ public class DeleteUsuarioServlet extends HttpServlet implements Servlet {
 	
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		Integer id = Integer.parseInt(req.getParameter("id"));
-		usuarioService.delete(id);
-		resp.sendRedirect("/tierramedia/usuario/lista.adm");
+		PerfilUsuario user = (PerfilUsuario) req.getSession().getAttribute("user");
+		if (user != null && user.isAdmin()) {
+			Integer id = Integer.parseInt(req.getParameter("id"));
+			usuarioService.delete(id);
+			resp.sendRedirect("/tierramedia/usuario/lista.adm");
+		} else {
+			resp.sendRedirect("/tierramedia/welcome");
+		}
 	}
 }
